feat(cli): add generate-data command (#1021)
<!-- markdownlint-disable -->

## Purpose

Sub-PR 4 of 6 for the [CLI RFC
#906](https://github.com/vllm-project/speculators/issues/906)
implementation (umbrella: #1024).

Supersedes #938 (old sequential topology).

- Move offline hidden-states generation from
`scripts/data_generation_offline.py` into
`speculators.cli.generate_data`
- Wire as `speculators generate-data` under the Pipeline help panel
- Original script becomes a deprecation shim (will be removed in v0.9.0)
- Verbose module docstring restored from original script with updated
CLI usage examples
- 4 smoke tests for the generate-data command (help, flags, rank
validation, concurrency validation)

Depends on: #1018 (CLI infrastructure)

### Migration diff

```bash
git diff main:scripts/data_generation_offline.py cli/generate-data:src/speculators/cli/generate_data.py
```

## Tests

```
$ pytest tests/unit/cli/ -v
tests/unit/cli/test_cli.py::TestRootApp::test_no_args_shows_help PASSED
tests/unit/cli/test_cli.py::TestRootApp::test_help PASSED
tests/unit/cli/test_cli.py::TestRootApp::test_version PASSED
tests/unit/cli/test_cli.py::TestRootApp::test_pipeline_commands_in_help PASSED
tests/unit/cli/test_cli.py::TestRootApp::test_tools_commands_in_help PASSED
tests/unit/cli/test_cli.py::TestConvertCommand::test_help PASSED
tests/unit/cli/test_cli.py::TestConvertCommand::test_algorithm_choices_in_help PASSED
tests/unit/cli/test_cli.py::TestConvertCommand::test_missing_required_args PASSED
tests/unit/cli/test_cli.py::TestGenerateDataCommand::test_help PASSED
tests/unit/cli/test_cli.py::TestGenerateDataCommand::test_fail_on_error_in_help PASSED
tests/unit/cli/test_cli.py::TestGenerateDataCommand::test_invalid_rank PASSED
tests/unit/cli/test_cli.py::TestGenerateDataCommand::test_invalid_concurrency PASSED
======================== 12 passed in 0.16s ========================
```

## Checklist

I have filled in:

- [x] The purpose of the PR, such as "Fix some issue (link existing
issues this PR will resolve)".
- [x] The test plan/results, such as providing test command and pasting
the results.
- [x] (Optional) The necessary documentation update.
- [x] I (a human) have written or reviewed the code in this pr to the
best of my ability.

---------
