+++ docs/.nav.yml
@@ -34,7 +34,7 @@ nav:
     - CLI Reference:
           - cli/index.md
           - prepare-data: cli/prepare_data.md
-          - data_generation_offline.py: cli/data_generation_offline.md
+          - generate-offline-data: cli/data_generation_offline.md
           - launch_vllm.py: cli/launch_vllm.md
           - train.py: cli/train.md
           - response_regeneration: cli/response_regeneration.md
