+++ src/speculators/cli/__init__.py
@@ -11,6 +11,7 @@ from importlib.metadata import version as pkg_version
 import typer
 
 from speculators.cli.convert import convert
+from speculators.cli.generate_offline_data import generate_offline_data
 from speculators.cli.prepare_data import prepare_data
 from speculators.cli.regenerate_responses import regenerate_responses
 from speculators.cli.stitch import stitch_command
@@ -44,6 +45,7 @@ def _main(
 
 app.command(rich_help_panel="Pipeline")(prepare_data)
 app.command(name="stitch-mtp", rich_help_panel="Pipeline")(stitch_command)
+app.command(rich_help_panel="Pipeline")(generate_offline_data)
 app.command(rich_help_panel="Pipeline")(regenerate_responses)
 app.command(
     name="train",
